requirements.txt
# Core AI/ML Dependencies
crewai>=1.8.1
crewai-tools>=0.1.6
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.34
langchain-anthropic>=0.1.0
openai>=1.14.2

# Vector Database & Embeddings
chromadb>=0.4.24
sentence-transformers>=2.5.1

# Web Framework & UI
streamlit>=1.32.2
gradio>=4.25.0

# Data Processing
pandas>=2.2.3
numpy>=1.26.4
python-dotenv>=1.0.1

# Web Scraping & Search
beautifulsoup4>=4.12.3
requests>=2.31.0
duckduckgo-search>=5.3.0

# Development & Testing
pytest>=8.1.1
pytest-cov>=4.1.0
black>=24.3.0
flake8>=7.0.0

# Deployment
docker>=7.0.0
gunicorn>=21.2.0

# Logging & Monitoring
python-json-logger==2.0.7

# Google Blogger/Blogspot Integration
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
markdown>=3.5.0

# Automation & Scheduling
schedule>=1.2.0

# Trending Topics Discovery
beautifulsoup4>=4.12.3  # Already included above
